TheoryProxy: Ensure that theory decisions are requested exhaustively at least once.
We have to ensure that if the SAT solver determines sat without making
any decisions, theory decisions are still requested until fixed point at
least once since some modules, e.g., finite model finding, rely on this.
MiniSat always called `pickBranchLit()` to make a decision at least
once, but this is not generally guaranteed in off-the-shelf,
state-of-the-art SAT solvers (including CaDiCaL). This is in preparation
for the CaDiCaL CDCL(T) integration.